Fixed constraints localiztion. Added Delegate tableVC to CellVC (not working). Extensions in their own folder. Removed repeated code in CellVC.
