+++ ColorsDescriptor/Extenstion/AddColorVCDelegateExtension.swift
@@ -1,0 +1,15 @@
+//
+//  AddColorVCDelegate.swift
+//  ColorsDescriptor
+//
+//  Created by Belal Hmeidat on 10/24/23.
+//
+
+import Foundation
+
+/// delegate extension to refresh the table data after adding a color action is performed in the AddColoViewController
+extension ColorsViewController : AddColorViewControllerDelegate {
+    func actionRequiresReloadPerformed() {
+        tableView.reloadData()
+    }
+}
